OBC Flight Software
* stop after first missing package

* implement uplinkAbortionReport (not integrated)

* add ST[1,6] to intermidiate steps and make stepId uint16

* mission test init

* add adcs events

* add 	message.data_size_message_ = message.data_size_ecss_ in TC and TM parsing

* increase sched TCs to 20

* add payload timer events

* finding the way out of tatsFs forest

* wip

* update Spacecraft id

* remove sequence number validation

* fix byte between bits error and wip lft

* clean up uplink

* set ECSSFunctionMaxArgLength = 175

* add explicit byte checks in intermidiate part uplink, fix total size calulation. IMPORTANT NOTE: fist intermidiate part must have sequnceNumber=0

* notify tc handling task with value 0

* add ST13 handling for firmware file updload directly to flash

* write to 2nd partion

* add crc log at the end of st13

* write to actual second partition

* add events

* Revert "add ST13 handling for firmware file updload directly to flash"

This reverts commit b684b44fa0f9020de5fb0ad980812469d604acb6.
And all relative flash commits

* make excpected value event unexcpected value event

* add events

* increase max scheduled tcs to 30

* refactor updateParameterFromMessage to get values as uint64 and handle internaly

* reduce max scheduled tcs to 25

* update max pmon definitions

* match ECSSEventDataAuxiliaryMaxSize to context size

+++ inc/Helpers/TypeDefinitions.hpp
@@ -1,7 +1,7 @@
 #pragma once
 #include <cstdint>
 
-using StepId = uint8_t;
+using StepId = uint16_t;
 using ParameterReportStructureId = uint16_t;
 using EventDefinitionId = uint16_t;
 using EventActionId = uint16_t;
